step 2 of job error handler refactor
make simple_error_handler top level

moves simple_error_handler into its own module so it can be imported without
causing import loops that happen if it lives alongside JobErrorHandler
